Add one a CSS animation folder
In this project , <div></div> is the element we’re animating. Looking at the CSS, we see that our animation declarations are associated with the div selector. The most important declaration here is animation-name, which binds the keyframe my-animation to our div element. Below are several additional declarations that affect the timing and behavior of the animation.

The animation itself is created with a keyframe, indicated by the @keyframes rule. A keyframe defines the animation’s starting state (inside from{}) and its end state (inside to{}). The keyframe my-animation changes three style properties of our div: background-color, width, and top. When these three properties are animated at once, it produces a coherent animation.

In this project, we only have one keyframe. Documents with multiple types of animations may have multiple keyframes, each bound to a different element.